+++ chain/transfer_tx_test.go
@@ -133,7 +133,7 @@ func TestTransferTx(t *testing.T) {
 		tc := &TransactionContext{
 			Genesis:   g,
 			Database:  db,
-			BlockTime: uint64(tv.blockTime),
+			BlockTime: tv.blockTime,
 			TxID:      ids.Empty,
 			Sender:    tv.sender,
 		}
